Simplify SyncDef and include stacks/checksums
As promised in the last commit, this simplifies SyncDef to just carry
the resources that need to be applied, with a bit of extra bookkeeping
in the form of checksums.

This commit also wires a flag through from the command line to the
cluster implementation, so one can switch garbage collection on. The
wiring bypasses the sync package, for simplicity -- it will always
calculate checksums.